Give a mob room for a boss past two billion HP
`max_hp` was an int32, which stops at 2.1B. The Chaos and Hard fights GMS
ships run to 84B, so the field they would be written into could not hold
them. Nothing narrows it on the way out -- the fight tracks HP as a double
and the detail panel already reads it as an int64.

+++ src/frontend/screens/boss_select_panel_test.cc
@@ -22,7 +22,7 @@
 namespace ms {
 namespace {
 
-Mob BossMob(const std::string& name, int level, int max_hp, int pdr) {
+Mob BossMob(const std::string& name, int level, int64_t max_hp, int pdr) {
   Mob mob;
   mob.set_name(name);
   mob.set_level(level);
